Ensure R directory is present before attempting to update package mirror config (#4332)

+++ templates/workspace_services/guacamole/user_resources/guacamole-azure-windowsvm/terraform/vm_config.ps1
@@ -53,9 +53,12 @@ local({
 "@
 
 $RBasePath = "$Env:ProgramFiles\R"
-$RVersions = Get-ChildItem -Path $RBasePath -Directory | Where-Object { $_.Name -like "R-*" }
 
-foreach ($RVersion in $RVersions) {
-    $ConfigPath = Join-Path -Path $RVersion.FullName -ChildPath "etc\Rprofile.site"
-    $RConfig | Out-File -Encoding Ascii (New-Item -Path $ConfigPath -Force)
+if (Test-Path $RBasePath) {
+  $RVersions = Get-ChildItem -Path $RBasePath -Directory | Where-Object { $_.Name -like "R-*" }
+
+  foreach ($RVersion in $RVersions) {
+      $ConfigPath = Join-Path -Path $RVersion.FullName -ChildPath "etc\Rprofile.site"
+      $RConfig | Out-File -Encoding Ascii (New-Item -Path $ConfigPath -Force)
+  }
 }
